+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.3.0
+
+* `query` now uses a queue instead of recursion
+* `query` should now correctly handle duplicate query detection
+* `dependencies_inflector` and `dependents_inflector` now handle inflection failures by using the original uninflected value
+* `dependencies_inflector` and `dependents_inflector` now use database metadata to determine primary key column names
+* `dependencies_inflector` and `dependents_inflector` now take an optional fourth parameter, `inflector`, to support custom inflection
+
 ## 0.2.1
 
 * Fixed issue that prevented `pycopyql.resolver` and its contents from being included in the release.